refactor(Tabs): separate agent and client with folders
