+++ packages/react/src/Tabs/Agent/Tabs.agent.mdx
