+++ packages/react/src/Tabs/Agent/components/Pane.tsx
