+++ packages/react/src/Tabs/Agent/components/TabsCore.tsx
