+++ packages/react/src/Tabs/Client/Tabs.client.mdx
