+++ packages/react/src/client.ts
@@ -1,25 +1,25 @@
 import "@fontsource/source-sans-pro";
 
+export { Accordion } from "./Accordion/Client/Accordion";
+export { Alert } from "./Alert/Alert.client";
+export type { AlertType } from "./Alert/Alert.client";
 export {
   ButtonClient as Button,
   Variants as ButtonVariants,
 } from "./Button/Button.client";
-export { Header } from "./Header/Header.client";
-export { RadioItem, Radio, RadioInput, RadioModes } from "./Form/Radio/Client";
+export { Footer } from "./Footer/Footer.client";
 export {
-  CheckboxItem,
   Checkbox,
   CheckboxInput,
+  CheckboxItem,
   CheckboxModes,
 } from "./Form/Checkbox/Client";
-export { Footer } from "./Footer/Footer.client";
+export { Radio, RadioInput, RadioItem, RadioModes } from "./Form/Radio/Client";
+export { Header } from "./Header/Header.client";
+export { Modal } from "./ModalClient/Modal.client";
+export type { ButtonAction as ModalButtonAction } from "./ModalClient/Modal.client";
+export { Stepper } from "./Stepper/Stepper.client";
 export {
   TabsClient as Tabs,
   Direction as TabsDirection,
-} from "./Tabs/Tabs.client";
-export { Stepper } from "./Stepper/Stepper.client";
-export { Accordion } from "./Accordion/Client/Accordion";
-export { Modal } from "./ModalClient/Modal.client";
-export type { ButtonAction as ModalButtonAction } from "./ModalClient/Modal.client";
-export { Alert } from "./Alert/Alert.client";
-export type { AlertType } from "./Alert/Alert.client";
+} from "./Tabs/Client/Tabs.client";
